chore: Format the repository using Prettier command (#16)
## Description
This PR does a bulk format to all the `.tsx`, `.ts`, `.jsx`, `.js`,
`.json`, `.yml` files in the repository.

Added this as a separate PR as bulk changes are expected. 

## Testing
Running the following command in the root repository to produce the file
changes make in this PR.
```bash
npx prettier . --write
````

## Follow-up Implementation
Add GitHub Actions for Prettier Formatting Checks to ensure that all
code in the repository are always formatted accordingly