feat: Add Formatting Checks to GitHub Actions (#19)
## Description
Currently, there are no formatting checks for the repo. To ensure the
code consistency of the code in the repo, we will be introducing
formatting checks in this PR.

As part of the integration, we will be adding a script as follows: 
For formatting checks, we can run the command below. 
```bash
npm run format:check
```

For formatting the code, we can run the command below. 
```bash
npm run format:check
```

This ensure consistency with the checks run in GitHub Actions and the
checks run manually.

## Things to Note
This PR should only be merged after the following PR is merged: #16.  
_Update: Already merged._