feat: refactor question service and add leetcode service (#24)
## Description
In this PR, we will work on refactoring and extending from another PR #9
which sets up the Question Service and basic function required to
communicate between LeetCode GraphQL to MongoDB.

## Main Changes

- [x] Refactor Question Service to Question and LeetCode Service for
better maintainability and scalability.
- [x] Enable batch transfer from LeetCode GraphQL to MongoDB
`leetcode-service` database
- [x] Automated POST to `question-service.questions` for every update in
`leetcode-service.questions`
- [x] Add GET status for status and random question for specified
`categoryTitle` and `difficulty`
- [x] Add Cron Job to run daily sync for questions form leetcode. 

## Service Responsibilities
### LeetCode Service

- Fetch questions from LeetCode GraphQL.
- Insert into `leetcode-service.questions`.
- Post added questions to question service

### Question Service

- Insert document into `questions-service.questions`.
- Endpoints to retrieve a specific question or aggregate status
(difficulty/category).

## API End points

### LeetCode Service
Base URL:
[http://localhost:5285/api/leetcode/](http://localhost:5285/api/leetcode/)
Batch update question:
[http://localhost:5285/api/leetcode/seed-batch](http://localhost:5285/api/leetcode/seed-batch)

### Question Service
Base URL:
[http://localhost:5275/api/question](http://localhost:5275/api/questions/)
Get Category and Difficulty:
[http://localhost:5275/api/questions/exists?categoryTitle={categoryTitle}&difficulty={difficulty}](http://localhost:5275/api/questions/exists?categoryTitle={categoryTitle}&difficulty={difficulty})
Get Random Question:
[http://localhost:5275/api/questions/random?categoryTitle={categoryTitle}&difficulty={difficulty}](http://localhost:5275/api/questions/random?categoryTitle={categoryTitle}&difficulty={difficulty})
Add question:
[http://localhost:5275/api/questions/post-question](http://localhost:5275/api/question/post-question)

## Rationale
Previously, the idea was to have a one-command bulk transfer of ALL
leetcode questions into the question bank in MongoDB.
But there is the issue of huge transfer load - since we need to transfer
the details of all 3000+ leetcode question. As such, this PR focuses on
the approach of batch updates (of 200 Question per transfer), where we
have an index that we keep track of and use that to track updates of the
update progress when done manually. The approach of batch transfers
could reduce DB load, simplify retries, and allow incremental progress
tracking.

## Further Extension / Edge Cases to be Considered
 - Seed Cursor registers edge cases: 
1. Docs with sync issues to `leetcode-service` but are not captured /
retired. Instead, they are being forgotten entirely.
2. Issue with sync from `leetcode-service` to `question-service` are not
captured and will be forgotten.
 - Additional queue to register for bad question: 
 - Expand test coverage.
 - Notifying for bad document during cron job

## Other Changes / Implementation
- Added `dbLimiter` to limit the access to the DB
- Add `logger.ts` for consistent logging for the services individually