refactor: clean docker file (#59)
## Description

This PR refactors the Dockerfiles for typescript services to implement a
multi-stage build. The Dockerfile for the `matching-service` is
excluded, as it has already been refactored to use a multi-stage build.
It also cleans up the Dockerfile for the Javascript services.

### Why Use Multi-Stage Builds?

- **Smaller Image Size**
By copying only the necessary files and dependencies into the final
image, the runtime environment is cleaned of any build dependencies,
reducing the overall image size. The build stage is discarded after it’s
done, keeping the final image lean and efficient.
- **Better Caching**
The Docker build process is optimized by copying the package.json files
before the rest of the application code. This allows Docker to cache the
layer that installs dependencies, avoiding redundant installs on
subsequent builds.
- **Separation of Concerns**
The build and runtime environments are clearly separated. The build
stage contains all dependencies needed to compile or bundle the app,
while the runtime stage is focused only on the final production setup.