Fix: Chat service not working in production (#99)
Description
---
This pull request adds a fallback option for chat service to store user
state management in the local instance if in the event redis adapter
cannot be initialised. The fallback solution does not support horizontal
scaling since it functions in the event redis adapter fails to
initialise, but allows us to demo the feature's functionality at least
when a single instance is running. This pull request also includes a
minor fix in user service.